Organize package and rename according to package naming convention
