--- src/com/java/javalava/game/MissionTemplate/Mission.java
@@ -1,5 +1,0 @@
-package com.java.javalava.game.MissionTemplate;
-
-public class Mission {
-
-}
