+++ src/com/javalava/game/mission_template/Mission.java
@@ -1,0 +1,5 @@
+package com.javalava.game.mission_template;
+
+public class Mission {
+
+}
